Split code around execution
